Added in Help Command
This commit adds in the help command all-beit a little hacky. Needless to say it's nice to be able to just tell someone to install this, and have the help command so they don't have to keep looking at the github page, or know how to read go. I mean honestly if they're using this they should be familiar enough to read strings but nonetheless.